fix issue : cannot work with webpack etc.
