+++ lib/MultiParty.js
@@ -4,16 +4,17 @@
  * SkyWayで簡単にマルチパーティ接続を実現するライブラリ
  */
 
+const EventEmitter = require('events').EventEmitter
+const util = require('./util')
 
-(function(global){
+navigator.getUserMedia_ = navigator.getUserMedia
+  || navigator.webkitGetUserMedia
+  || navigator.mozGetUserMedia;
 
-  navigator.getUserMedia_ = navigator.getUserMedia
-    || navigator.webkitGetUserMedia
-    || navigator.mozGetUserMedia;
-  window.AudioContext = window.AudioContext || window.webkitAudioContext;
+class MultiParty extends EventEmitter {
+  constructor(opts) {
+    super();
 
-  // コンストラクタ
-  var MultiParty_ = function(opts){
     this.room = null; // ルーム名
     this.id = null; // ID
     this.key = null; // app key
@@ -26,18 +27,11 @@
 
     // option をチェック
     // room, myid, key プロパティを割り当てる
-    this.opts = MultiParty_.util.checkOpts_(opts);
+    this.opts = util.checkOpts_(opts);
   }
 
-  // EventEmitterを継承する
-  MultiParty_.prototype = new EventEmitter2();
-
-  ///////////////////////////////////
-  // private method
-
-
   // SkyWayサーバーに繋ぐ
-  MultiParty_.prototype.conn2SkyWay_ = function() {
+  conn2SkyWay_() {
     var self = this;
 
 
@@ -72,10 +66,10 @@
     this.peer.on("error", function(err) {
       self.fire_('error', err)
     });
- }
+  }
 
   // 接続中のIDを取得する
-  MultiParty_.prototype.getIDs_ = function() {
+  getIDs_() {
     var self = this;
 
     self.listAllPeers(function(peers){
@@ -94,7 +88,7 @@
   }
 
   // 接続中のIDとサーバのIDを比較して再接続・切断する
-  MultiParty_.prototype.startPollingConnections_ = function(){
+  startPollingConnections_(){
     var self = this;
     self.pollInterval = setInterval(function(){
       self.listAllPeers(function(newList){
@@ -140,12 +134,12 @@
   // video 処理
 
   // Media Stream 処理の開始
-  MultiParty_.prototype.startMediaStream_ = function(){
+  startMediaStream_(){
     this.startMyStream_();
   }
 
   // video stream 取得開始
-  MultiParty_.prototype.startMyStream_ = function() {
+  startMyStream_() {
     var self = this;
 
     navigator.getUserMedia_({"video": self.opts.video_stream, "audio": self.opts.audio_stream},
@@ -176,7 +170,7 @@
   // ```
   // multiparty.mute({audio: true});  // only audio becomes mute, so callee becomes not to hear about caller voice.
   // ```
-  MultiParty_.prototype.mute = function(target_track /* {"video": boolean, "audio": boolean} */) {
+  mute(target_track /* {"video": boolean, "audio": boolean} */) {
     // if parameter **target_track** does not proper value. We forcibly set both mute.
     if( typeof(target_track) !== "object" ) { target_track = {video:true, audio:true}; }
 
@@ -192,7 +186,7 @@
   // ```
   // multiparty.unmute({audio: true});  // only audio becomes unmute, so caller's face cannot be seen from callee side.
   // ```
-  MultiParty_.prototype.unmute = function(target_track /* {"video": boolean, "audio": boolean} */) {
+  unmute(target_track /* {"video": boolean, "audio": boolean} */) {
     // if parameter **target_track** does not proper value. We forcibly set both unmute.
     if( typeof(target_track) !== "object") { target_track = {video:true, audio:true}; }
 
@@ -203,7 +197,7 @@
 
 
   // peersに対して、MediaStream callを開始する
-  MultiParty_.prototype.startCall_ = function(isScreen) {
+  startCall_(isScreen) {
     var self = this;
 
     // API経由で取得したIDには、自分からcallする（caller）
@@ -262,7 +256,7 @@
   }
 
   // peerからのvideo stream, closeに対し、ハンドラをセットする
-  MultiParty_.prototype.setupStreamHandler_ = function(call) {
+  setupStreamHandler_(call) {
     var self = this;
     var isReconnect = !!(call.metadata && call.metadata.reconnect);
 
@@ -318,7 +312,7 @@
   }
 
   // peerのvideoのObjectURLを生成し、frontにpeer_msイベントを返す
-  MultiParty_.prototype.setupPeerVideo_ = function(peer_id, stream, isReconnect) {
+  setupPeerVideo_(peer_id, stream, isReconnect) {
     // prevent to call twice.
     // if(!!this.peers[peer_id].video) return;
 
@@ -333,7 +327,7 @@
 
   // peerのvideo Nodeをセットアップする
   // loadedmetadataが完了したら、'peer_video'をfireする
-  MultiParty_.prototype.setupPeerScreen_ = function(peer_id, stream, isReconnect) {
+  setupPeerScreen_(peer_id, stream, isReconnect) {
     var self = this;
     if(!isReconnect){
       isReconnect = false;
@@ -344,7 +338,7 @@
   }
 
   // peerのdcとmcを全てクローズする
-  MultiParty_.prototype.removePeer = function(peer_id) {
+  removePeer(peer_id) {
     try{
       if(this.peers[peer_id] !== undefined) {
         var peer = this.peers[peer_id];
@@ -371,12 +365,12 @@
   // DataChannel 処理
 
   // DataChannel 処理を開始する
-  MultiParty_.prototype.startDataChannel_ = function(){
+  startDataChannel_(){
     this.startDCconnection_();
   }
 
   // DataChannelのコネクション処理を行う
-  MultiParty_.prototype.startDCconnection_ = function(){
+  startDCconnection_(){
     var self = this;
 
     // API経由で取得したIDには、自分からconnectする
@@ -397,7 +391,7 @@
   }
 
   // DataChannelのコネクション処理を行う
-  MultiParty_.prototype.DCconnect_ = function(peer_id){
+  DCconnect_(peer_id){
     var conn = this.peer.connect(peer_id, {"serialization": this.opts.serialization, "reliable": this.opts.reliable});
     this.peers[peer_id].DCconn = conn;
 
@@ -408,7 +402,7 @@
   }
 
   // DataChannelのイベントハンドラをセットする
-  MultiParty_.prototype.setupDCHandler_ = function(conn) {
+  setupDCHandler_(conn) {
     var self = this;
     conn.on('data', function(data) {
       self.fire_('message', {"id": this.peer, "data": data});
@@ -416,7 +410,6 @@
       // handle peer close event
       // check skyway server to see this user is disconnected.
       var peer_id = this.peer;
-      var metadata = this.metadata;
       self.listAllPeers(function(list) {
         var isDisconnected = true;
         for (var index in list) {
@@ -442,7 +435,7 @@
   }
 
   // DataChannelでつながっているpeerにメッセージを送信する
-  MultiParty_.prototype.send_ = function(data) {
+  send_(data) {
     if(!this.peer) {
       return false;
     }
@@ -459,163 +452,14 @@
   }
 
   // イベントを発火する
-  MultiParty_.prototype.fire_ = function(name, obj) {
+  fire_(name, obj) {
     this.emit(name, obj);
   }
-
-
-
-  /////////////////////////////////////
-  // utility
-  MultiParty_.util = {};
-
-  // オプションのチェック
-  MultiParty_.util.checkOpts_ = function(opts_) {
-    var opts = {};
-
-    // key check (なかったら throw)
-    if(!opts_.key || typeof(opts_.key) !== "string") {
-      throw "app key must be specified";
-    };
-
-    // key check ( string patter がマッチしなかったら throw )
-    if(!opts_.key.match(/^[0-9a-z]{8}\-[0-9a-z]{4}\-[0-9a-z]{4}\-[0-9a-z]{4}\-[0-9a-z]{12}$/)) {
-      throw "wrong string pattern of app key";
-    };
-   // copy key
-    opts.key = opts_.key;
-
-    // todo : room prefix にdomainを意識したほげほげ
-    // room check (なかったら "")
-    if(!opts_.room || typeof(opts_.room) !== "string") {
-      var seed = "";
-    } else if(!opts_.room.match(/^[0-9a-zA-Z\-\_]{4,32}$/)){
-      throw "room name should be digit|alphabet and length between 4 and 32";
-    } else {
-      var seed = opts_.room
-    };
-
-    opts.room_name = seed;
-
-    opts.room_id = MultiParty_.util.makeRoomName(seed);
-
-    // id check (なかったら生成）
-
-    var hash_ = location.pathname + "_" + opts.room_id + "_peer_id";
-    if(!!sessionStorage[hash_]) {
-      opts.id = sessionStorage[hash_];
-    } else if(!opts_.id || typeof(opts_.id) !== "string") {
-      opts.id = opts.room_id + MultiParty_.util.makeID();
-    } else {
-      opts.id = opts.room_id + opts_.id;
-    }
-    sessionStorage[hash_] = opts.id;
-
-    // reliable check (なかったら false)
-    if(!opts_.reliable) {
-      opts.reliable = false;
-    } else {
-      opts.reliable = true;
-    }
-
-    // serialization check (未指定なら binary)
-    if(!opts_.serialization) {
-      opts.serialization = "binary";
-    } else {
-      // serializationのタイプをチェックする
-      // binary, utf-8, json以外はエラー
-      opts.serialization = opts_.serialization;
-    }
-
-    // stream check
-    opts.video_stream = (opts_.video === undefined ? true : opts_.video);
-    opts.audio_stream = (opts_.audio === undefined ? true : opts_.audio);
-    opts.use_stream = opts.video_stream || opts.audio_stream;
-
-    // polling disconnect/reconnect (must for FF)
-    opts.polling = (opts_.polling === undefined ? true : opts_.polling);
-    opts.polling_interval = (opts_.polling_interval === undefined ? 3000 : opts_.polling_interval);
-
-    // peerjs options
-    opts.peerjs_opts = {
-      debug: opts_.debug || false,
-      key: opts.key
-    };
-    if(opts_.host){
-      opts.peerjs_opts.host = opts_.host;
-    }
-    if(opts_.port){
-        opts.peerjs_opts.port = opts_.port;
-    }
-    if(opts_.secure){
-      opts.peerjs_opts.secure = opts_.secure;
-    }
-    if(opts_.config){
-      opts.peerjs_opts.config = opts_.config;
-    }
-
-    return opts;
-  }
-
-
-
-  // IDを作る
-  MultiParty_.util.makeID = function() {
-    var id = "";
-
-    for (var i = 0; i < 32; i++) {
-      // id += String.fromCharCode( (Math.random() * (125 - 33) + 33) | 0)
-      id += String.fromCharCode( (Math.random() * (57 - 48) + 48) | 0)
-    }
-    return id;
-  }
-
-  // room名を作る
-  MultiParty_.util.makeRoomName = function(seed) {
-      seed = seed || "";
-      seed += location.host + location.pathname;
-      return CybozuLabs.MD5.calc(seed).substring(0,6) + "R_";
-  }
-
-  // video nodeを作る
-  MultiParty_.util.createVideoNode = function(video) {
-    // 古いノードが会った場合は削除
-    var prev_node = document.getElementById(video.id);
-    if(prev_node) prev_node.parentNode.removeChild(prev_node);
-
-    // 表示用のビデオノードを作る
-    var v_ = document.createElement("video");
-    v_.setAttribute("src", video.src);
-    v_.setAttribute("id", video.id);
-
-    var played = false;
-
-    v_.addEventListener("loadedmetadata", function(ev) {
-      if(!played) {
-        played = true;
-        this.play();
-      }
-    }, false);
-
-    // since FF37 sometimes doesn't fire "loadedmetadata"
-    // work around is after 500msec, calling play();
-    setTimeout(function(ev){
-      if(!played) {
-        played = true;
-        v_.play();
-      }
-    }, 500);
-
-    return v_;
-  }
-
-
-
 
   ////////////////////////////////////
   // public method
 
-  MultiParty_.prototype.start = function() {
+  start() {
     var self = this;
     setTimeout(function(){
       self.conn2SkyWay_();
@@ -623,18 +467,18 @@
     return self;
   }
 
-  MultiParty_.prototype.send = function(data) {
+  send(data) {
     if(this.peer) this.send_(data);
   }
 
   // 切断する
-  MultiParty_.prototype.close = function() {
+  close() {
     if(this.peer) this.peer.destroy();
     clearInterval(this.pollInterval);
   }
 
   // 画面共有を開始する
-  MultiParty_.prototype.startScreenShare = function(success, error) {
+  startScreenShare(success, error) {
       if(this.peer) {
           var self = this;
           if(SkyWay && SkyWay.ScreenShare) {
@@ -659,7 +503,7 @@
 
 
   // 画面共有を停止する
-  MultiParty_.prototype.stopScreenShare = function() {
+  stopScreenShare() {
     if(this.screenStream){
       this.screenStream.getVideoTracks()[0].stop();
       for(var peer_id in this.peers){
@@ -673,7 +517,7 @@
   }
 
   // 同じRoomのpeer listを取得
-  MultiParty_.prototype.listAllPeers = function(callback) {
+  listAllPeers(callback) {
     var self = this;
     this.peer.listAllPeers(
       function(peers){
@@ -689,7 +533,7 @@
   }
 
   // ユーザに再接続する
-  MultiParty_.prototype.reconnect = function(peer_id, connections) {
+  reconnect(peer_id, connections) {
     var self = this;
     var peer = self.peers[peer_id];
     if(connections === undefined) {
@@ -745,15 +589,9 @@
       }
     }
   }
+}
 
+MultiParty.util = util;
 
-
-
-
-  // オブジェクトの宣言
-  if (!global.MultiParty) {
-    global.MultiParty = MultiParty_;
-  }
-}(window));
-
+module.exports = MultiParty;
