+++ lib/index.js
@@ -1,0 +1,7 @@
+const MultiParty = require("./MultiParty")
+
+if(module.require) {
+  module.exports = MultiParty;
+} else {
+  window.MultiParty = MultiParty;
+}
